Deprecate zklogin sdk, migrate to sui sdk
