+++ apps/wallet/package.json
@@ -122,7 +122,6 @@
 		"@mysten/ledgerjs-hw-app-sui": "workspace:*",
 		"@mysten/sui.js": "workspace:*",
 		"@mysten/wallet-standard": "workspace:*",
-		"@mysten/zklogin": "workspace:*",
 		"@noble/hashes": "^1.3.1",
 		"@radix-ui/react-checkbox": "^1.0.4",
 		"@radix-ui/react-collapsible": "^1.0.3",
