+++ sdk/docs/pages/typescript/index.mdx
@@ -93,5 +93,4 @@ what you need to keep your code light and compact.
 - [`@mysten/sui.js/utils`](/typescript/utils) - Utilities for formatting and parsing various Sui
   types.
 - [`@mysten/sui.js/faucet`](/typescript/faucet) - Methods for requesting SUI from a faucet.
-- [`@mysten/sui.js/zklogin`](/typescript/zklogin) - Utilities for working with zkLogin. (Note
-  @mysten/zklogin package also contains zklogin related functionality)
+- [`@mysten/sui.js/zklogin`](/typescript/zklogin) - Utilities for working with zkLogin.
