+++ sdk/docs/pages/typescript/zklogin.mdx
@@ -1,13 +1,7 @@
-import { Callout } from 'nextra/components';
-
 # ZkLogin
 
 Utilities for working with zlLogin. Currently contains functionality to create and parse zkLogin
 signatures and compute zkLogin addresses.
-
-<Callout type="info">
-	Note: `@mysten/zklogin` package contains more utilities to help with computing address etc.
-</Callout>
 
 To parse a serialized zkLogin signature
