+++ sdk/typescript/package.json
@@ -131,6 +131,8 @@
 		"@scure/bip32": "^1.3.1",
 		"@scure/bip39": "^1.2.1",
 		"@suchipi/femver": "^1.0.0",
+		"jose": "^4.14.4",
+		"poseidon-lite": "^0.2.0",
 		"superstruct": "^1.0.3",
 		"tweetnacl": "^1.0.3"
 	}
