+++ sdk/typescript/src/zklogin/index.ts
@@ -2,7 +2,19 @@
 // SPDX-License-Identifier: Apache-2.0
 
 export { getZkLoginSignature, parseZkLoginSignature } from './signature.js';
-export { toBigEndianBytes } from './utils.js';
-export { computeZkLoginAddressFromSeed } from './address.js';
 export { ZkLoginPublicIdentifier } from './publickey.js';
 export type { ZkLoginSignatureInputs } from './bcs.js';
+
+export { computeZkLoginAddress, jwtToAddress, computeZkLoginAddressFromSeed } from './address.js';
+export type { ComputeZkLoginAddressOptions } from './address.js';
+
+export { poseidonHash } from './poseidon.js';
+
+export { generateNonce, generateRandomness } from './nonce.js';
+
+export {
+	hashASCIIStrToField,
+	genAddressSeed,
+	getExtendedEphemeralPublicKey,
+	toBigEndianBytes,
+} from './utils.js';
