+++ sdk/typescript/test/unit/zklogin/nonce.test.ts
@@ -3,8 +3,8 @@
 
 import { expect, test } from 'vitest';
 
-import { Ed25519Keypair } from '../../typescript/src/keypairs/ed25519';
-import { generateNonce, generateRandomness } from '../src';
+import { Ed25519Keypair } from '../../../src/keypairs/ed25519';
+import { generateNonce, generateRandomness } from '../../../src/zklogin';
 
 test('can generate using `generateRandomness`', () => {
 	const kp = Ed25519Keypair.fromSecretKey(new Uint8Array(32));
