--- sdk/zklogin/.prettierignore
@@ -1,2 +1,0 @@
-CHANGELOG.md
-dist/
