--- sdk/zklogin/README.md
@@ -1,6 +1,0 @@
-# `@mysten/zklogin`
-
-> **⚠️ This package is experimental and will change rapidly as it is being developed. Do not
-> consider these APIs to be stable. If you have any feedback,
-> [open an issue](https://github.com/MystenLabs/sui/issues/new/choose) or message us on
-> [Discord](https://discord.gg/Sui).**
