--- sdk/zklogin/package.json
@@ -1,56 +1,0 @@
-{
-	"name": "@mysten/zklogin",
-	"version": "0.4.1",
-	"description": "Utilities for interacting with zkLogin in Sui",
-	"license": "Apache-2.0",
-	"author": "Mysten Labs <build@mystenlabs.com>",
-	"type": "commonjs",
-	"main": "./dist/cjs/index.js",
-	"module": "./dist/esm/index.js",
-	"types": "./dist/cjs/index.d.ts",
-	"exports": {
-		".": {
-			"import": "./dist/esm/index.js",
-			"require": "./dist/cjs/index.js"
-		}
-	},
-	"sideEffects": false,
-	"files": [
-		"CHANGELOG.md",
-		"dist",
-		"src"
-	],
-	"scripts": {
-		"clean": "rm -rf tsconfig.tsbuildinfo ./dist",
-		"build": "build-package",
-		"prepublishOnly": "pnpm build",
-		"prettier:check": "prettier -c --ignore-unknown .",
-		"prettier:fix": "prettier -w --ignore-unknown .",
-		"eslint:check": "eslint --max-warnings=0 .",
-		"eslint:fix": "pnpm run eslint:check --fix",
-		"lint": "pnpm run eslint:check && pnpm run prettier:check",
-		"lint:fix": "pnpm run eslint:fix && pnpm run prettier:fix",
-		"test": "vitest"
-	},
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/mystenlabs/sui.git"
-	},
-	"bugs": {
-		"url": "https://github.com/mystenlabs/sui/issues"
-	},
-	"homepage": "https://github.com/mystenlabs/sui#readme",
-	"devDependencies": {
-		"@mysten/build-scripts": "workspace:*",
-		"@types/node": "^20.4.2",
-		"typescript": "^5.1.6",
-		"vitest": "^0.33.0"
-	},
-	"dependencies": {
-		"@mysten/bcs": "workspace:*",
-		"@mysten/sui.js": "workspace:*",
-		"@noble/hashes": "^1.3.1",
-		"jose": "^4.14.4",
-		"poseidon-lite": "^0.2.0"
-	}
-}
