--- sdk/zklogin/src/index.ts
@@ -1,13 +1,0 @@
-// Copyright (c) Mysten Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-export { computeZkLoginAddress, jwtToAddress } from './address.js';
-export type { ComputeZkLoginAddressOptions } from './address.js';
-
-export { getZkLoginSignature } from '@mysten/sui.js/zklogin';
-
-export { poseidonHash } from './poseidon.js';
-
-export { generateNonce, generateRandomness } from './nonce.js';
-
-export { hashASCIIStrToField, genAddressSeed, getExtendedEphemeralPublicKey } from './utils.js';
