--- sdk/zklogin/tsconfig.esm.json
@@ -1,7 +1,0 @@
-{
-	"extends": "./tsconfig.json",
-	"compilerOptions": {
-		"module": "ESNext",
-		"outDir": "dist/esm"
-	}
-}
