--- sdk/zklogin/tsconfig.json
@@ -1,11 +1,0 @@
-{
-	"extends": "../build-scripts/tsconfig.shared.json",
-	"include": ["src"],
-	"compilerOptions": {
-		"module": "CommonJS",
-		"outDir": "dist/cjs",
-		"isolatedModules": true,
-		"rootDir": "src"
-	},
-	"references": [{ "path": "../typescript" }, { "path": "../bcs" }]
-}
